requirements.txt
flask>-1.1
gunicorn>=0.14.2
pynacl>=1.4.0
watchdog>=0.8.2
networkx>=2.5
